# Improve `Oj.dump` performance with `Time` object