@ohler55 2021-08-29
Two comments to address first. The first is easy. The second may require some re-thinking.