@ohler55 2021-08-30
Not a waste of time at all. It was nice working with you.